email_classification_project-5/requirements.txt

flask
pyyaml
torch
transformers
datasets
pymongo
werkzeug
